+++ src/radical/pilot/agent/radical-pilot-agent-multicore.py
@@ -620,10 +620,13 @@ class Scheduler(threading.Thread):
         rpu.prof('schedule', msg="allocated", uid=cu['_id'], logger=self._log.warn)
         self._log.info (self.slot_status())
 
-        cu_list = rpu.blowup(self._config, cu, EXECUTION_QUEUE)
+        cu_list, cu_dropped = rpu.blowup(self._config, cu, EXECUTION_QUEUE)
         for _cu in cu_list :
             rpu.prof('put', msg="Scheduler to execution_queue (%s)" % _cu['state'], uid=_cu['_id'])
             self._execution_queue.put(_cu)
+
+        # we need to free allocated cores for dropped CUs
+        self.unschedule(cu_dropped)
 
         return True
 
@@ -692,7 +695,7 @@ class Scheduler(threading.Thread):
 
                 request = self._schedule_queue.get()
 
-                if not isinstance(list, request):
+                if not isinstance(request, list):
                     # command only, no cu
                     request = [request, None]
 
@@ -706,8 +709,17 @@ class Scheduler(threading.Thread):
 
                 rpu.prof('get_cmd', msg="schedule_queue to Scheduler (%s)" % command)
 
-                if command == COMMAND_RESCHEDULE:
+                if command == COMMAND_WAKEUP:
 
+                    # nothing to do (other then testing self._terminate)
+                    rpu.prof('get_cmd', msg="schedule_queue to Scheduler (wakeup)")
+                    continue
+
+
+                elif command == COMMAND_RESCHEDULE:
+
+                    # reschedule is done over all units in the wait queue
+                    assert (cu == None) 
                     self._reschedule()
 
 
@@ -719,7 +731,8 @@ class Scheduler(threading.Thread):
                     cu['state'] = rp.ALLOCATING
 
 
-                    cu_list  = rpu.blowup(self._config, cu, SCHEDULER)
+
+                    cu_list, _  = rpu.blowup(self._config, cu, SCHEDULER)
                     for _cu in cu_list:
 
                         # we got a new unit to schedule.  Either we can place it
@@ -733,8 +746,6 @@ class Scheduler(threading.Thread):
 
 
                 elif command == COMMAND_UNSCHEDULE :
-
-                    cu = data
 
                     # we got a finished unit, and can re-use its cores
                     #
@@ -847,8 +858,7 @@ class SchedulerContinuous(Scheduler):
 
         # TODO: single_node should be enforced for e.g. non-message passing
         #       tasks, but we don't have that info here.
-        # NOTE AM: why should non-messaging tasks be confined to one node?
-        if cores_requested < self._lrms.cores_per_node:
+        if cores_requested <= self._lrms.cores_per_node:
             single_node = True
         else:
             single_node = False
@@ -1783,7 +1793,6 @@ class LaunchMethodDPLACE(LaunchMethod):
         return dplace_command, None
 
 
-
 # ==============================================================================
 #
 class LaunchMethodMPIRUNRSH(LaunchMethod):
@@ -1794,13 +1803,18 @@ class LaunchMethodMPIRUNRSH(LaunchMethod):
 
         LaunchMethod.__init__(self, name, config, logger, scheduler)
 
-
     # --------------------------------------------------------------------------
     #
     def _configure(self):
-        # mpirun_rsh (e.g. on Gordon@ SDSC)
-        self.launch_command = self._which('mpirun_rsh')
 
+        # mpirun_rsh (e.g. on Gordon@SDSC, Stampede@TACC)
+        if not self._which('mpirun_rsh'):
+            raise Exception("mpirun_rsh could not be found")
+
+        # We don't use the full pathname as the user might load a different
+        # compiler / MPI library suite from his CU pre_exec that requires
+        # the launcher from that version, as experienced on stampede in #572.
+        self.launch_command = 'mpirun_rsh'
 
     # --------------------------------------------------------------------------
     #
@@ -3270,7 +3284,7 @@ class ExecWorker(COMPONENT_TYPE):
 
     # --------------------------------------------------------------------------
     #
-    def __init__(self, name, config, logger, agent, lrms, scheduler,
+    def __init__(self, name, config, logger, agent, scheduler,
                  task_launcher, mpi_launcher, command_queue,
                  execution_queue, stageout_queue, update_queue, 
                  schedule_queue, pilot_id, session_id):
@@ -3284,7 +3298,6 @@ class ExecWorker(COMPONENT_TYPE):
         self._config           = config
         self._log              = logger
         self._agent            = agent
-        self._lrms             = lrms
         self._scheduler        = scheduler
         self._task_launcher    = task_launcher
         self._mpi_launcher     = mpi_launcher
@@ -3304,7 +3317,7 @@ class ExecWorker(COMPONENT_TYPE):
     # This class-method creates the appropriate sub-class for the Launch Method.
     #
     @classmethod
-    def create(cls, name, config, logger, spawner, agent, lrms, scheduler,
+    def create(cls, name, config, logger, spawner, agent, scheduler,
                task_launcher, mpi_launcher, command_queue,
                execution_queue, update_queue, schedule_queue, 
                stageout_queue, pilot_id, session_id):
@@ -3319,7 +3332,7 @@ class ExecWorker(COMPONENT_TYPE):
                 SPAWNER_NAME_SHELL : ExecWorker_SHELL
             }[spawner]
 
-            impl = implementation(name, config, logger, agent, lrms, scheduler,
+            impl = implementation(name, config, logger, agent, scheduler,
                                   task_launcher, mpi_launcher, command_queue,
                                   execution_queue, stageout_queue, update_queue, 
                                   schedule_queue, pilot_id, session_id)
@@ -3372,7 +3385,7 @@ class ExecWorker_POPEN (ExecWorker) :
 
     # --------------------------------------------------------------------------
     #
-    def __init__(self, name, config, logger, agent, lrms, scheduler,
+    def __init__(self, name, config, logger, agent, scheduler,
                  task_launcher, mpi_launcher, command_queue,
                  execution_queue, stageout_queue, update_queue, 
                  schedule_queue, pilot_id, session_id):
@@ -3385,7 +3398,7 @@ class ExecWorker_POPEN (ExecWorker) :
         self._cu_environment = self._populate_cu_environment()
 
 
-        ExecWorker.__init__ (self, name, config, logger, agent, lrms, scheduler,
+        ExecWorker.__init__ (self, name, config, logger, agent, scheduler,
                  task_launcher, mpi_launcher, command_queue,
                  execution_queue, stageout_queue, update_queue, 
                  schedule_queue, pilot_id, session_id)
@@ -3462,7 +3475,7 @@ class ExecWorker_POPEN (ExecWorker) :
 
                 try:
 
-                    cu_list = rpu.blowup(self._config, cu, EXEC_WORKER)
+                    cu_list, _ = rpu.blowup(self._config, cu, EXEC_WORKER)
                     for _cu in cu_list:
 
                         if _cu['description']['mpi']:
@@ -3631,7 +3644,7 @@ class ExecWorker_POPEN (ExecWorker) :
                                       state  = rp.EXECUTING,
                                       msg    = "unit execution start")
 
-        cu_list = rpu.blowup(self._config, cu, WATCH_QUEUE)
+        cu_list, _ = rpu.blowup(self._config, cu, WATCH_QUEUE)
         for _cu in cu_list :
             rpu.prof('put', msg="ExecWorker to watcher (%s)" % _cu['state'], uid=_cu['_id'])
             self._watch_queue.put(_cu)
@@ -3686,7 +3699,7 @@ class ExecWorker_POPEN (ExecWorker) :
                 for cu in cus :
                     
                     rpu.prof('get', msg="ExecWatcher picked up unit", uid=cu['_id'])
-                    cu_list = rpu.blowup(self._config, cu, WATCHER)
+                    cu_list, _ = rpu.blowup(self._config, cu, WATCHER)
 
                     for _cu in cu_list :
                         self._cus_to_watch.append (_cu)
@@ -3779,7 +3792,7 @@ class ExecWorker_POPEN (ExecWorker) :
                                                   state  = rp.STAGING_OUTPUT,
                                                   msg    = "unit execution completed")
 
-                    cu_list = rpu.blowup(self._config, cu, STAGEOUT_QUEUE)
+                    cu_list, _ = rpu.blowup(self._config, cu, STAGEOUT_QUEUE)
                     for _cu in cu_list :
                         rpu.prof('put', msg="ExecWatcher to stageout_queue (%s)" % _cu['state'], uid=_cu['_id'])
                         self._stageout_queue.put(_cu)
@@ -3794,12 +3807,14 @@ class ExecWorker_SHELL(ExecWorker):
 
     # --------------------------------------------------------------------------
     #
-    def __init__(self, name, config, logger, agent, lrms, scheduler,
+    def __init__(self, name, config, logger, agent, scheduler,
                  task_launcher, mpi_launcher, command_queue,
                  execution_queue, stageout_queue, update_queue,
                  schedule_queue, pilot_id, session_id):
 
-        ExecWorker.__init__ (self, name, config, logger, agent, lrms, scheduler,
+        rpu.prof('ExecWorker init')
+
+        ExecWorker.__init__ (self, name, config, logger, agent, scheduler,
                  task_launcher, mpi_launcher, command_queue,
                  execution_queue, stageout_queue, update_queue,
                  schedule_queue, pilot_id, session_id)
@@ -3893,7 +3908,7 @@ class ExecWorker_SHELL(ExecWorker):
 
                 try:
 
-                    cu_list = rpu.blowup(self._config, cu, EXEC_WORKER)
+                    cu_list, _ = rpu.blowup(self._config, cu, EXEC_WORKER)
 
                     for _cu in cu_list :
 
@@ -3956,6 +3971,9 @@ class ExecWorker_SHELL(ExecWorker):
 
             ret = list()
             for arg in args :
+
+                if not arg:
+                    continue
 
                 # if string is between outer single quotes,
                 #    pass it as is.
@@ -4263,7 +4281,7 @@ class ExecWorker_SHELL(ExecWorker):
                                           state = rp.STAGING_OUTPUT,
                                           msg   = "unit execution completed")
 
-            cu_list = rpu.blowup(self._config, cu, STAGEOUT_QUEUE)
+            cu_list, _ = rpu.blowup(self._config, cu, STAGEOUT_QUEUE)
 
             for _cu in cu_list :
                 rpu.prof('put', msg="ExecWatcher to stageout_queue (%s)" % _cu['state'], uid=_cu['_id'])
@@ -4384,7 +4402,7 @@ class UpdateWorker(threading.Thread):
                 else:
                     rpu.prof('get', msg="update_queue to UpdateWorker", uid=uid)
 
-                update_request_list = rpu.blowup(self._config, update_request, UPDATE_WORKER)
+                update_request_list, _ = rpu.blowup(self._config, update_request, UPDATE_WORKER)
 
                 for _update_request in update_request_list :
 
@@ -4479,7 +4497,7 @@ class StageinWorker(threading.Thread):
 
                 rpu.prof('get', msg="stagein_queue to StageinWorker (%s)" % cu['state'], uid=cu['_id'])
 
-                cu_list = rpu.blowup(self._config, cu, STAGEIN_WORKER)
+                cu_list, _ = rpu.blowup(self._config, cu, STAGEIN_WORKER)
                 for _cu in cu_list :
 
                     _cu['state'] = rp.STAGING_INPUT
@@ -4589,7 +4607,7 @@ class StageinWorker(threading.Thread):
                                                       state  = rp.ALLOCATING,
                                                       msg    = 'agent input staging done')
 
-                        _cu_list = rpu.blowup(self._config, _cu, SCHEDULE_QUEUE)
+                        _cu_list, _ = rpu.blowup(self._config, _cu, SCHEDULE_QUEUE)
                         for __cu in _cu_list :
                             rpu.prof('put', msg="StageinWorker to schedule_queue (%s)" % __cu['state'], uid=__cu['_id'])
                             self._schedule_queue.put([COMMAND_SCHEDULE, __cu])
@@ -4670,7 +4688,7 @@ class StageoutWorker(threading.Thread):
 
                 rpu.prof('get', msg="stageout_queue to StageoutWorker (%s)" % cu['state'], uid=cu['_id'])
 
-                cu_list = rpu.blowup(self._config, cu, STAGEOUT_WORKER)
+                cu_list, _ = rpu.blowup(self._config, cu, STAGEOUT_WORKER)
                 for _cu in cu_list :
 
                     sandbox = os.path.join(self._workdir, '%s' % _cu['_id'])
@@ -5095,7 +5113,6 @@ class Agent(object):
                 spawner         = spawner,
                 logger          = self._log,
                 agent           = self,
-                lrms            = self._lrms,
                 scheduler       = self._scheduler,
                 task_launcher   = self._task_launcher,
                 mpi_launcher    = self._mpi_launcher,
@@ -5195,7 +5212,7 @@ class Agent(object):
         else:
             rpu.prof('put', msg="%s to update_queue" % src, uid=query_dict['_id'])
 
-        query_list = rpu.blowup(self._config, query_dict, UPDATE_QUEUE)
+        query_list, _ = rpu.blowup(self._config, query_dict, UPDATE_QUEUE)
 
         for _query_dict in query_list :
             self._update_queue.put({'_id'    : _query_dict['_id'],
@@ -5383,7 +5400,7 @@ class Agent(object):
 
             rpu.prof('get', msg="MongoDB to Agent (%s)" % cu['state'], uid=cu['_id'], logger=self._log.info)
 
-            _cu_list = rpu.blowup(self._config, cu, AGENT)
+            _cu_list, _ = rpu.blowup(self._config, cu, AGENT)
             for _cu in _cu_list :
 
                 try:
@@ -5421,7 +5438,7 @@ class Agent(object):
                                                state  = rp.STAGING_INPUT,
                                                msg    = 'unit needs input staging')
 
-                        _cu_list = rpu.blowup(self._config, _cu, STAGEIN_QUEUE)
+                        _cu_list, _ = rpu.blowup(self._config, _cu, STAGEIN_QUEUE)
                         for __cu in _cu_list :
                             rpu.prof('put', msg="Agent to stagein_queue (%s)" % __cu['state'], uid=__cu['_id'])
                             self._stagein_queue.put(__cu)
@@ -5433,7 +5450,7 @@ class Agent(object):
                                                state  = rp.ALLOCATING,
                                                msg    = 'unit needs no input staging')
 
-                        _cu_list = rpu.blowup(self._config, _cu, SCHEDULE_QUEUE)
+                        _cu_list, _ = rpu.blowup(self._config, _cu, SCHEDULE_QUEUE)
                         for __cu in _cu_list :
                             rpu.prof('put', msg="Agent to schedule_queue (%s)" % __cu['state'], uid=__cu['_id'])
                             self._schedule_queue.put([COMMAND_SCHEDULE, __cu])
@@ -5487,7 +5504,7 @@ def main():
 
     if args : parser.error("Unused arguments '%s'" % args)
 
-    if not options.cores                : parser.error("Missing number of cores (-c)")
+    if not options.cores                : parser.error("Missing or zero number of cores (-c)")
     if not options.debug_level          : parser.error("Missing DEBUG level (-d)")
     if not options.task_launch_method   : parser.error("Missing unit launch method (-j)")
     if not options.mpi_launch_method    : parser.error("Missing mpi launch method (-k)")
@@ -5497,10 +5514,10 @@ def main():
     if not options.spawner              : parser.error("Missing agent spawner (-o)")
     if not options.pilot_id             : parser.error("Missing pilot id (-p)")
     if not options.agent_scheduler      : parser.error("Missing agent scheduler (-q)")
-    if not options.runtime              : parser.error("Missing agent runtime (-r)")
+    if not options.runtime              : parser.error("Missing or zero agent runtime (-r)")
     if not options.session_id           : parser.error("Missing session id (-s)")
 
-    rpu.prof('start', uid=options.pilot_id)
+    rpu.prof_init('agent.prof', 'start', uid=options.pilot_id)
 
     # configure the agent logger
     logger    = logging.getLogger  ('radical.pilot.agent')